+++ src/idx_search_kappa.cpp
@@ -246,11 +246,11 @@ static void idx_search_thread_body(IdxSearchShared *S)
 			for (uint j = 0; j < n; ++j)
 				{
 				const uint Nbr = NeighborKmers[j];
-				const uint RowSize = Index.GetRowSize(Nbr);
+				const uint64_t RowSize = Index.GetRowSize(Nbr);
 				if (RowSize == 0)
 					continue;
-				uint DataOffset = Index.GetRowStart(Nbr);
-				for (uint c = 0; c < RowSize; ++c)
+				uint64_t DataOffset = Index.GetRowStart(Nbr);
+				for (uint64_t c = 0; c < RowSize; ++c)
 					{
 					uint32_t TSeqIdx;
 					uint16_t TPos;
